Duplicate Nash Equilibriums
These functions take in the already generated training samples with nash equilibriums and duplicate them until there are 10 nash equilibrium in each game sample for the model. Then it stores the duplicated nash equilibriums in an other CSV file for convenience. The ideal location to call the functions is after line 45 or  trainingSamples = np.array(trainingSamples) and before line 51. I also sent in the old and new nash equilibrium for comparison